# docs/swagger.yaml
definitions:
  models.ErrorDTO:
    properties:
      error:
        type: string
    required:
    - error
    type: object
  models.GroupRegistInput:
    properties:
      category:
        type: string
      description:
        type: string
      name:
        type: string
      permission:
        type: boolean
    required:
    - category
    - description
    - name
    type: object
  models.InfoDTO:
    properties:
      info:
        type: string
    required:
    - info
    type: object
  models.UserSignupInput:
    properties:
      contact:
        type: string
      email:
        type: string
      name:
        type: string
      nickname:
        type: string
      password:
        type: string
      passwordCheck:
        type: string
    required:
    - email
    - password
    - passwordCheck
    type: object
info:
  contact: {}
paths:
  /group/regist:
    post:
      consumes:
      - application/json
      description: 그룹 등록 API
      operationId: group-regist
      parameters:
      - description: 그룹 등록을 위한 양식
        in: body
        name: newGroupForm
        required: true
        schema:
          $ref: '#/definitions/models.GroupRegistInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InfoDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorDTO'
      summary: 그룹 등록
  /user/signup:
    post:
      consumes:
      - application/json
      description: 유저 회원가입 API
      operationId: user-signup
      parameters:
      - description: 유저 회원가입을 위한 양식
        in: body
        name: newUserForm
        required: true
        schema:
          $ref: '#/definitions/models.UserSignupInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InfoDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorDTO'
      summary: 유저 회원가입
swagger: "2.0"